#29 - Corrigindo warning 'conversion to B from A may change the sign of the result' da compilação do mud

+++ MUDGameProgramming/Libraries/SocketLib/Connection.h
@@ -236,7 +236,9 @@ void Connection<protocol>::BufferData(const char *p_buffer, size_t p_size) {
 template <class protocol> void Connection<protocol>::SendBuffer() {
   if (m_sendbuffer.size() > 0) {
     // send the data, and erase as much as was sent from the buffer.
-    int sent = Send(m_sendbuffer.data(), static_cast<int>(m_sendbuffer.size()));
+    size_t sent = static_cast<size_t>(
+      Send(m_sendbuffer.data(), m_sendbuffer.size())
+    );
     m_sendbuffer.erase(0, sent);
 
     if (sent > 0) {
@@ -268,7 +270,7 @@ template <class protocol> void Connection<protocol>::SendBuffer() {
 // ------------------------------------------------------------------------
 template <class protocol> void Connection<protocol>::Receive() {
   // receive the data
-  int bytes = DataSocket::Receive(m_buffer, BUFFERSIZE);
+  ssize_t bytes = DataSocket::Receive(m_buffer, BUFFERSIZE);
 
   // get the current time
   BasicLib::sint64 t = BasicLib::GetTimeS();
